Use explicit data type when binding.
Using explicit data types helped me prevent errors when binding to a boolean column, without using explicit data types ```false``` can become ```""``` and have the query fail. This was tested in postgres. Optionally I could have cast to an integer when setting the bind value, but that shouldn't have to be the case.