Redo 3.x Parser to be more like 2.x Rebuilder
This PR converts the @jlacoude lexer parser logic to something like the 2.x rebuilder logic. It is not as full-featured as the lexer, but it is much faster. The bundled driver-specific parsers do not handle comments, but custom parsers can be injected for those cases.

As with the @jlacoude work, multiple named placeholders of the same name are allowed; likewise, numbered placeholders are converted to named placeholders. The renaming style has changed, so that repeated names are suffixed with `__` instead of `_`, and start at 1 instead of 0; likewise, the conversion of numbered placeholders to names now uses the name `__` instead of `__numbered`.

Tests for comments and multiple queries have been removed, as the new logic does not support them. Also removed is the PgsqlParserTest::testUnicodeStringConstants(), as I don't see how to tell the new parser logic to ignore colons uses as Unicode escapers in Postgres. Finally, the testEscapedCharactersInStringConstants() `SELECT "Escaping \\\\" :foo ""` is commented out, as it fails. Otherwise, all the pre-existing @jlacoude tests pass, though I have reorganized them under a base abstract test to remove duplication.

Finally, the benchmarks on this are much closer to the 2.x benchmarks than previous attempts.

$ php bench-2.x.php;php bench-2.x.php;php bench-2.x.php;php bench-2.x.php;php bench-2.x.php;
Total: 0.24769711494446
Total: 0.23618698120117
Total: 0.2351438999176
Total: 0.23142695426941
Total: 0.23610711097717
$ php bench-3.x.php;php bench-3.x.php;php bench-3.x.php;php bench-3.x.php;php bench-3.x.php;
Total: 0.25684905052185
Total: 0.26446294784546
Total: 0.25762104988098
Total: 0.27490496635437
Total: 0.27310609817505

This comes from the tradeoff of not handling comments and other cases.

@pavarnos @jlacoude @harikt please criticize as harshly as you can.